+++ src/rovr/functions/config.py
@@ -4,6 +4,7 @@ from importlib import resources
 from importlib.metadata import PackageNotFoundError, version
 from os import environ, path
 from platform import system
+from shutil import which
 
 import jsonschema
 import toml
@@ -343,6 +344,19 @@ def load_config() -> tuple[dict, dict]:
         config["plugins"]["editor"]["folder_executable"] = environ.get(
             "EDITOR", "vim" if system() != "Windows" else "code"
         )
+    # pdf fixer
+    if (
+        config["plugins"]["poppler"]["enabled"]
+        and config["plugins"]["poppler"]["poppler_folder"] == ""
+    ):
+        pdfinfo_path = path.dirname(which("pdfinfo"))  # type: ignore[no-matching-overload]
+        if pdfinfo_path is None:
+            pprint(
+                "[WARN] Poppler is enabled, but no poppler folder was specified, and it was not found in PATH. "
+                "[WARN] Please install Poppler and set the poppler_folder in rovr config.",
+                style="yellow",
+            )
+        config["plugins"]["poppler"]["poppler_folder"] = pdfinfo_path
     return schema, config